+++ Text/Megaparsec/Error.hs
@@ -41,6 +41,7 @@ import Data.Set (Set)
 import Data.Typeable (Typeable)
 import GHC.Generics
 import Prelude hiding (concat)
+import Test.QuickCheck hiding (label)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Set           as E
 
@@ -62,6 +63,18 @@ data ErrorItem t
   deriving (Show, Read, Eq, Ord, Data, Typeable, Generic)
 
 instance NFData t => NFData (ErrorItem t)
+
+instance Arbitrary t => Arbitrary (ErrorItem t) where
+  arbitrary = oneof
+    [
+#if !MIN_VERSION_QuickCheck(2,9,0)
+      Tokens <$> (NE.fromList . getNonEmpty <$> arbitrary)
+    , Label  <$> (NE.fromList . getNonEmpty <$> arbitrary)
+#else
+      Tokens <$> arbitrary
+    , Label  <$> arbitrary
+#endif
+    , return EndOfInput ]
 
 -- | The type class defines how to represent information about various
 -- exceptional situations. Data types that are used as custom data component
@@ -103,6 +116,13 @@ data Dec
 instance NFData Dec where
   rnf (DecFail str) = rnf str
   rnf (DecIndentation ord ref act) = ord `seq` rnf ref `seq` rnf act
+
+instance Arbitrary Dec where
+  arbitrary = oneof
+    [ sized (\n -> do
+        k <- choose (0, n `div` 2)
+        DecFail <$> vectorOf k arbitrary)
+    , DecIndentation <$> arbitrary <*> arbitrary <*> arbitrary ]
 
 instance ErrorComponent Dec where
   representFail        = DecFail
@@ -151,6 +171,19 @@ instance ( Show t
 #if MIN_VERSION_base(4,8,0)
   displayException = parseErrorPretty
 #endif
+
+instance (Arbitrary t, Ord t, Arbitrary e, Ord e)
+    => Arbitrary (ParseError t e) where
+  arbitrary = ParseError
+    <$>
+#if !MIN_VERSION_QuickCheck(2,9,0)
+      (NE.fromList . getNonEmpty <$> arbitrary)
+#else
+      arbitrary
+#endif
+    <*> arbitrary
+    <*> arbitrary
+    <*> arbitrary
 
 -- | Merge two error data structures into one joining their collections of
 -- message items and preferring longest match. In other words, earlier error